Move RejectionFeedbackSurveyComponent to CandidateInterface
  This component will only be used by candidates
