+++ app/components/candidate_interface/rejection_reasons/rejection_feedback_survey_component.html.erb
